[zshrc] Reimplement newbeamer as function
newbeamer now takes two arguments, namely: `folder` and `name`.
`folder` is where the presentation will be stored at. (default: `.`)
`name` is the name of the folder containing the presentation template.
(default: `Presentation`)

It will check if the `folder/name` exists and exit if it does. If the
`folder/name` does not exist, it will copy the template to that path.